+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## UNRELEASED
+
+### Changed
+- Querying for `raw` data in preview to allow retrieving files above 1MB size from the repo using the API
+
 ## `0.0.1`
 
 Initial version